feat: do not raise MissingAccountData
The MissingAccountData error is supposed
to be a developer-caused error. While this
may indeed be true, it provides no clear
path forward for the user, whereas a re-auth
would typically solve the issue.

This commit replaces this error in the internal
AccountInfo implementation with InvalidAuthToken
which should force a re-auth